requirements.txt
google-api-python-client
google-cloud
google-cloud-storage
google-cloud-vision